Update resume
I had a resume that included my time at Amazon Seattle but I hadn't linked to it from my blog. This change refreshes that link. The resume doesn't include my time in Tokyo though.

Also using S3 to manage the PDF to separate the responsibility of managing blobs versus code. Not exactly sure what to do with the images, which are still managed by git.